Lumen Updates
Updated for Lumen functionality
