+++ lumen/.env
@@ -10,7 +10,7 @@ LOG_SLACK_WEBHOOK_URL=
 
 DB_CONNECTION=mysql
 DB_HOST=127.0.0.1
-DB_PORT=8889
+DB_PORT=3306
 DB_DATABASE=pokemon_db
 DB_USERNAME=root
 DB_PASSWORD=
