+++ lumen/routes/web.php
@@ -19,3 +19,4 @@ $router->get('/', function () use ($router) {
 
 $router->get("/lines", "LineController@getAll");
 $router->get("/lines/{id}", "LineController@getOne");
+$router->get("/trainer-lines", "TrainerLineController@getAll");
